refactor service module
